[ig-scorer/refactor-main-scorer] merge in AI tests
