--- packages/perseus-score/src/widgets/interactive-graph/score-interactive-graph-backup.test.ts
@@ -1,1252 +1,0 @@
-// STOPSHIP: don't land this as-is, it needs to be removed
-
-import invariant from "tiny-invariant";
-import _ from "underscore";
-
-import scoreInteractiveGraph from "./score-interactive-graph";
-
-import type {
-    PerseusGraphType,
-    PerseusInteractiveGraphRubric,
-} from "@khanacademy/perseus-core";
-
-describe("InteractiveGraph scoring on a segment question", () => {
-    it("marks the answer invalid if guess is undefined", () => {
-        const guess = undefined;
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {
-                type: "segment",
-            },
-            correct: {
-                type: "segment",
-                coords: [
-                    [
-                        [0, 0],
-                        [1, 1],
-                    ],
-                ],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("marks the answer invalid if guess.coords is missing", () => {
-        const guess: PerseusGraphType = {type: "segment"};
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {
-                type: "segment",
-            },
-            correct: {
-                type: "segment",
-                coords: [
-                    [
-                        [0, 0],
-                        [1, 1],
-                    ],
-                ],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("does not award points if guess.coords is wrong", () => {
-        const guess: PerseusGraphType = {
-            type: "segment",
-            coords: [
-                [
-                    [99, 0],
-                    [1, 1],
-                ],
-            ],
-        };
-
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "segment"},
-            correct: {
-                type: "segment",
-                coords: [
-                    [
-                        [0, 0],
-                        [1, 1],
-                    ],
-                ],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-
-    it("awards points if guess.coords is right", () => {
-        const guess: PerseusGraphType = {
-            type: "segment",
-            coords: [
-                [
-                    [0, 0],
-                    [1, 1],
-                ],
-            ],
-        };
-
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "segment"},
-            correct: {
-                type: "segment",
-                coords: [
-                    [
-                        [0, 0],
-                        [1, 1],
-                    ],
-                ],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("allows points of a segment to be specified in reverse order", () => {
-        const guess: PerseusGraphType = {
-            type: "segment",
-            coords: [
-                [
-                    [1, 1],
-                    [0, 0],
-                ],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "segment"},
-            correct: {
-                type: "segment",
-                coords: [
-                    [
-                        [0, 0],
-                        [1, 1],
-                    ],
-                ],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("does not modify the `guess` data", () => {
-        const guess: PerseusGraphType = {
-            type: "segment",
-            coords: [
-                [
-                    [1, 1],
-                    [0, 0],
-                ],
-            ],
-        };
-
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "segment"},
-            correct: {
-                type: "segment",
-                coords: [
-                    [
-                        [0, 0],
-                        [1, 1],
-                    ],
-                ],
-            },
-        };
-
-        scoreInteractiveGraph(guess, rubric);
-
-        expect(guess.coords).toEqual([
-            [
-                [1, 1],
-                [0, 0],
-            ],
-        ]);
-    });
-
-    it("does not modify `rubric`", () => {
-        const guess: PerseusGraphType = {
-            type: "segment",
-            coords: [
-                [
-                    [1, 1],
-                    [0, 0],
-                ],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "segment"},
-            correct: {
-                type: "segment",
-                coords: [
-                    [
-                        [1, 1],
-                        [0, 0],
-                    ],
-                ],
-            },
-        };
-
-        scoreInteractiveGraph(guess, rubric);
-
-        // Narrow the type of `rubric.correct` to segment graph; otherwise TS
-        // thinks it might not have a `coords` property.
-        invariant(rubric.correct.type === "segment");
-        expect(rubric.correct.coords).toEqual([
-            [
-                [1, 1],
-                [0, 0],
-            ],
-        ]);
-    });
-});
-
-describe("InteractiveGraph scoring on an angle question", () => {
-    it("marks the answer invalid if guess.coords is missing", () => {
-        const guess: PerseusGraphType = {type: "angle"};
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "angle"},
-            correct: {
-                type: "angle",
-                coords: [
-                    [1, 1],
-                    [0, 0],
-                    [-1, -1],
-                ],
-                allowReflexAngles: false,
-                match: "congruent",
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveInvalidInput();
-    });
-});
-
-describe("InteractiveGraph scoring on a point question", () => {
-    it("marks the answer invalid if guess.coords is missing", () => {
-        const guess: PerseusGraphType = {type: "point"};
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "point"},
-            correct: {
-                type: "point",
-                coords: [[0, 0]],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("throws an exception if correct.coords is missing", () => {
-        // Characterization test: this might not be desirable behavior, but
-        // it's the current behavior as of 2024-09-25.
-        const guess: PerseusGraphType = {
-            type: "point",
-            coords: [[0, 0]],
-        };
-
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {
-                type: "point",
-            },
-            // @ts-expect-error: Testing exception for invalid rubric
-            correct: {
-                type: "point",
-            },
-        };
-
-        expect(() => scoreInteractiveGraph(guess, rubric)).toThrow();
-    });
-
-    it("does not award points if guess.coords is wrong", () => {
-        const guess: PerseusGraphType = {
-            type: "point",
-            coords: [[9, 9]],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "point"},
-            correct: {
-                type: "point",
-                coords: [[0, 0]],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-
-    it("awards points if guess.coords is right", () => {
-        const guess: PerseusGraphType = {
-            type: "point",
-            coords: [[7, 8]],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "point"},
-            correct: {
-                type: "point",
-                coords: [[7, 8]],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("allows points to be specified in any order", () => {
-        const guess: PerseusGraphType = {
-            type: "point",
-            coords: [
-                [7, 8],
-                [5, 6],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "point"},
-            correct: {
-                type: "point",
-                coords: [
-                    [5, 6],
-                    [7, 8],
-                ],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("does not modify the `guess` data", () => {
-        const guess: PerseusGraphType = {
-            type: "point",
-            coords: [
-                [7, 8],
-                [5, 6],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "point"},
-            correct: {
-                type: "point",
-                coords: [
-                    [5, 6],
-                    [7, 8],
-                ],
-            },
-        };
-
-        const guessClone = _.clone(guess);
-
-        scoreInteractiveGraph(guess, rubric);
-
-        expect(guess).toEqual(guessClone);
-    });
-
-    it("does not modify `rubric`", () => {
-        const guess: PerseusGraphType = {
-            type: "point",
-            coords: [
-                [7, 8],
-                [5, 6],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "point"},
-            correct: {
-                type: "point",
-                coords: [
-                    [5, 6],
-                    [7, 8],
-                ],
-            },
-        };
-
-        const rubricClone = _.clone(rubric);
-
-        scoreInteractiveGraph(guess, rubric);
-
-        expect(rubric).toEqual(rubricClone);
-    });
-});
-
-describe("InteractiveGraph scoring on an angle question", () => {
-    it("marks the answer invalid if guess.coords is missing", () => {
-        // Arrange
-        const guess: PerseusGraphType = {type: "angle"};
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {
-                type: "angle",
-            },
-            correct: {
-                type: "angle",
-                allowReflexAngles: false,
-                coords: [
-                    [-5, 0],
-                    [0, 0],
-                    [5, 5],
-                ],
-            },
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        // Assert
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("properly reverses coordinates if angle graph is reflexive when not allowed", () => {
-        // Arrange
-        const guess: PerseusGraphType = {
-            type: "angle",
-            coords: [
-                [-5, 0],
-                [0, 0],
-                [5, 5],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {
-                type: "angle",
-            },
-            correct: {
-                type: "angle",
-                allowReflexAngles: false,
-                coords: [
-                    [5, 5],
-                    [0, 0],
-                    [-5, 0],
-                ],
-            },
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("does not reverse coordinates when angle is not reflexive", () => {
-        // Arrange
-        const guess: PerseusGraphType = {
-            type: "angle",
-            coords: [
-                [5, 0],
-                [0, 0],
-                [5, 5],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {
-                type: "angle",
-            },
-            correct: {
-                type: "angle",
-                allowReflexAngles: false,
-                coords: [
-                    [5, 0],
-                    [0, 0],
-                    [5, 5],
-                ],
-            },
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("does not reverse coordinates if the angle graph is allowed to be reflexive", () => {
-        // Arrange
-        const guess: PerseusGraphType = {
-            type: "angle",
-            coords: [
-                [5, 0],
-                [0, 0],
-                [-5, -5],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {
-                type: "angle",
-                coords: [
-                    [5, 0],
-                    [0, 0],
-                    [-5, -5],
-                ],
-            },
-            correct: {
-                type: "angle",
-                allowReflexAngles: true,
-                coords: [
-                    [5, 0],
-                    [0, 0],
-                    [-5, -5],
-                ],
-            },
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-});
-
-// Test data: f(x) = 2·4^x + 3  →  f(0)=5, f(1)=11, asymptote y=3
-const exponentialRubric: PerseusInteractiveGraphRubric = {
-    graph: {type: "exponential"},
-    correct: {
-        type: "exponential",
-        coords: [
-            [0, 5],
-            [1, 11],
-        ],
-        asymptote: 3,
-    },
-};
-
-describe("InteractiveGraph scoring on an exponential question", () => {
-    it("marks the answer invalid if guess is undefined", () => {
-        // Arrange, Act
-        const result = scoreInteractiveGraph(undefined, exponentialRubric);
-
-        // Assert
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("marks the answer invalid if coords are missing", () => {
-        // Arrange
-        const guess: PerseusGraphType = {type: "exponential"};
-
-        // Act
-        const result = scoreInteractiveGraph(guess, exponentialRubric);
-
-        // Assert
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("marks the answer incorrect if asymptote is missing", () => {
-        // Arrange — coords present (hasValue=true) but no asymptote,
-        // so the exponential scoring block is skipped and falls through to incorrect
-        const guess: PerseusGraphType = {
-            type: "exponential",
-            coords: [
-                [0, 5],
-                [1, 11],
-            ],
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, exponentialRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-
-    it("marks a correct answer as correct", () => {
-        // Arrange
-        const guess: PerseusGraphType = {
-            type: "exponential",
-            coords: [
-                [0, 5],
-                [1, 11],
-            ],
-            asymptote: 3,
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, exponentialRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("marks a wrong answer as incorrect", () => {
-        // Arrange — different curve: f(x) = 4·(1/4)^x, asymptote y=0
-        const guess: PerseusGraphType = {
-            type: "exponential",
-            coords: [
-                [0, 4],
-                [1, 1],
-            ],
-            asymptote: 0,
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, exponentialRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-
-    it("marks a correct answer as correct when coords are in reverse order", () => {
-        // Arrange — same curve, points swapped
-        const guess: PerseusGraphType = {
-            type: "exponential",
-            coords: [
-                [1, 11],
-                [0, 5],
-            ],
-            asymptote: 3,
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, exponentialRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-});
-
-const logarithmRubric: PerseusInteractiveGraphRubric = {
-    graph: {type: "logarithm"},
-    correct: {
-        type: "logarithm",
-        coords: [
-            [-4, -3],
-            [-5, -7],
-        ],
-        asymptote: -6,
-    },
-};
-
-describe("InteractiveGraph scoring on a logarithm question", () => {
-    it("marks the answer invalid if guess is undefined", () => {
-        // Arrange, Act
-        const result = scoreInteractiveGraph(undefined, logarithmRubric);
-
-        // Assert
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("marks the answer invalid if coords are missing", () => {
-        // Arrange
-        const guess: PerseusGraphType = {type: "logarithm"};
-
-        // Act
-        const result = scoreInteractiveGraph(guess, logarithmRubric);
-
-        // Assert
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("marks the answer incorrect if asymptote is missing", () => {
-        // Arrange — coords present (hasValue=true) but no asymptote,
-        // so the logarithm scoring block is skipped and falls through to incorrect
-        const guess: PerseusGraphType = {
-            type: "logarithm",
-            coords: [
-                [-4, -3],
-                [-5, -7],
-            ],
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, logarithmRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-
-    it("marks a correct answer as correct", () => {
-        // Arrange
-        const guess: PerseusGraphType = {
-            type: "logarithm",
-            coords: [
-                [-4, -3],
-                [-5, -7],
-            ],
-            asymptote: -6,
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, logarithmRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("marks a wrong answer as incorrect", () => {
-        // Arrange — different curve
-        const guess: PerseusGraphType = {
-            type: "logarithm",
-            coords: [
-                [2, 1],
-                [3, 2],
-            ],
-            asymptote: 0,
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, logarithmRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-
-    it("marks equivalent curves with different control points as correct", () => {
-        // Arrange — y = ln(x): both sets of points produce a=1, b=1, c=0
-        const lnRubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "logarithm"},
-            correct: {
-                type: "logarithm",
-                coords: [
-                    [1, 0],
-                    [Math.E, 1],
-                ],
-                asymptote: 0,
-            },
-        };
-        const guess: PerseusGraphType = {
-            type: "logarithm",
-            coords: [
-                [Math.E * Math.E, 2],
-                [Math.E * Math.E * Math.E, 3],
-            ],
-            asymptote: 0,
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, lnRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-});
-
-describe("InteractiveGraph scoring on an absolute-value question", () => {
-    it("marks the answer invalid if guess is undefined", () => {
-        // Arrange
-        const guess = undefined;
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "absolute-value"},
-            correct: {
-                type: "absolute-value",
-                coords: [
-                    [0, 0],
-                    [1, 1],
-                ],
-            },
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        // Assert
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("marks the answer invalid if guess.coords is missing", () => {
-        // Arrange
-        const guess: PerseusGraphType = {type: "absolute-value"};
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "absolute-value"},
-            correct: {
-                type: "absolute-value",
-                coords: [
-                    [0, 0],
-                    [1, 1],
-                ],
-            },
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        // Assert
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("does not award points if guess.coords is wrong", () => {
-        // Arrange
-        const guess: PerseusGraphType = {
-            type: "absolute-value",
-            coords: [
-                [0, 0],
-                [1, 2],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "absolute-value"},
-            correct: {
-                type: "absolute-value",
-                coords: [
-                    [0, 0],
-                    [1, 1],
-                ],
-            },
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-
-    it("awards points if guess.coords produce the same coefficients as the rubric", () => {
-        // Arrange
-        const guess: PerseusGraphType = {
-            type: "absolute-value",
-            coords: [
-                [0, 0],
-                [1, 1],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "absolute-value"},
-            correct: {
-                type: "absolute-value",
-                coords: [
-                    [0, 0],
-                    [1, 1],
-                ],
-            },
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("awards points when p2 is on the other arm but slope is the same", () => {
-        // Arrange — vertex at (0,0), slope=1 regardless of which side p2 is on
-        const guess: PerseusGraphType = {
-            type: "absolute-value",
-            coords: [
-                [0, 0],
-                [-1, 1],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "absolute-value"},
-            correct: {
-                type: "absolute-value",
-                coords: [
-                    [0, 0],
-                    [1, 1],
-                ],
-            },
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("does not award points for a downward-opening graph when upward is correct", () => {
-        // Arrange
-        const guess: PerseusGraphType = {
-            type: "absolute-value",
-            coords: [
-                [0, 0],
-                [1, -1],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "absolute-value"},
-            correct: {
-                type: "absolute-value",
-                coords: [
-                    [0, 0],
-                    [1, 1],
-                ],
-            },
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-
-    it("does not award points when p2 shares the same x as the vertex", () => {
-        // Arrange
-        const guess: PerseusGraphType = {
-            type: "absolute-value",
-            coords: [
-                [0, 0],
-                [0, 2], // same x as vertex → getAbsoluteValueCoefficients returns undefined
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "absolute-value"},
-            correct: {
-                type: "absolute-value",
-                coords: [
-                    [0, 0],
-                    [1, 1],
-                ],
-            },
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-});
-
-describe("InteractiveGraph scoring on a tangent question", () => {
-    it("marks the answer invalid if guess is undefined", () => {
-        const guess = undefined;
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "tangent"},
-            correct: {
-                type: "tangent",
-                coords: [
-                    [0.5, 0.5],
-                    [0.75, 0.75],
-                ],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("marks the answer invalid if guess has no coords", () => {
-        const guess: PerseusGraphType = {type: "tangent"};
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "tangent"},
-            correct: {
-                type: "tangent",
-                coords: [
-                    [0.5, 0.5],
-                    [0.75, 0.75],
-                ],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("scores correct answer as correct", () => {
-        const guess: PerseusGraphType = {
-            type: "tangent",
-            coords: [
-                [0.5, 0.5],
-                [0.75, 0.75],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "tangent"},
-            correct: {
-                type: "tangent",
-                coords: [
-                    [0.5, 0.5],
-                    [0.75, 0.75],
-                ],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("scores incorrect answer as incorrect", () => {
-        const guess: PerseusGraphType = {
-            type: "tangent",
-            coords: [
-                [1, 1],
-                [2, 3],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "tangent"},
-            correct: {
-                type: "tangent",
-                coords: [
-                    [0.5, 0.5],
-                    [0.75, 0.75],
-                ],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-
-    it("scores equivalent tangent curves as correct (different control points, same curve)", () => {
-        // Both sets of coords produce the same canonical tangent curve
-        // f(x) = tan(π/4 * x) — just shifted by one full period
-        const guess: PerseusGraphType = {
-            type: "tangent",
-            coords: [
-                [4, 0],
-                [5, 1],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "tangent"},
-            correct: {
-                type: "tangent",
-                coords: [
-                    [0, 0],
-                    [1, 1],
-                ],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("scores tangent with negative amplitude correctly", () => {
-        // Guess has coords that produce a negative amplitude tangent
-        const guess: PerseusGraphType = {
-            type: "tangent",
-            coords: [
-                [0, 0],
-                [1, -2],
-            ],
-        };
-        const rubric: PerseusInteractiveGraphRubric = {
-            graph: {type: "tangent"},
-            correct: {
-                type: "tangent",
-                coords: [
-                    [0, 0],
-                    [1, -2],
-                ],
-            },
-        };
-
-        const result = scoreInteractiveGraph(guess, rubric);
-
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-});
-
-const vectorRubric: PerseusInteractiveGraphRubric = {
-    graph: {type: "vector"},
-    correct: {
-        type: "vector",
-        coords: [
-            [0, 0],
-            [3, 4],
-        ],
-    },
-};
-
-describe("InteractiveGraph scoring on a vector question", () => {
-    it("marks the answer invalid if guess is undefined", () => {
-        // Arrange, Act
-        const result = scoreInteractiveGraph(undefined, vectorRubric);
-
-        // Assert
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("marks the answer invalid if coords are missing", () => {
-        // Arrange
-        const guess: PerseusGraphType = {type: "vector"};
-
-        // Act
-        const result = scoreInteractiveGraph(guess, vectorRubric);
-
-        // Assert
-        expect(result).toHaveInvalidInput();
-    });
-
-    it("marks a correct answer as correct", () => {
-        // Arrange
-        const guess: PerseusGraphType = {
-            type: "vector",
-            coords: [
-                [0, 0],
-                [3, 4],
-            ],
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, vectorRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("marks a wrong tail as incorrect", () => {
-        // Arrange — correct tip but wrong tail
-        const guess: PerseusGraphType = {
-            type: "vector",
-            coords: [
-                [1, 1],
-                [3, 4],
-            ],
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, vectorRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-
-    it("marks a wrong tip as incorrect", () => {
-        // Arrange — correct tail but wrong tip
-        const guess: PerseusGraphType = {
-            type: "vector",
-            coords: [
-                [0, 0],
-                [4, 5],
-            ],
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, vectorRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-
-    it("marks swapped tail and tip as incorrect", () => {
-        // Arrange — coords are reversed (tip at tail position, tail at tip)
-        // Unlike ray, vector scoring is order-sensitive
-        const guess: PerseusGraphType = {
-            type: "vector",
-            coords: [
-                [3, 4],
-                [0, 0],
-            ],
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, vectorRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-});
-
-const vectorCongruentRubric: PerseusInteractiveGraphRubric = {
-    graph: {type: "vector"},
-    correct: {
-        type: "vector",
-        coords: [
-            [0, 0],
-            [3, 4],
-        ],
-        match: "congruent",
-    },
-};
-
-describe("InteractiveGraph scoring on a vector question with congruent matching", () => {
-    it("marks a translated vector as correct when match is congruent", () => {
-        // Arrange — same direction ⟨3, 4⟩ but different position
-        const guess: PerseusGraphType = {
-            type: "vector",
-            coords: [
-                [2, 1],
-                [5, 5],
-            ],
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, vectorCongruentRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("marks exact position as correct when match is congruent", () => {
-        // Arrange — same position as rubric
-        const guess: PerseusGraphType = {
-            type: "vector",
-            coords: [
-                [0, 0],
-                [3, 4],
-            ],
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, vectorCongruentRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("marks wrong direction as incorrect when match is congruent", () => {
-        // Arrange — same magnitude but different direction
-        const guess: PerseusGraphType = {
-            type: "vector",
-            coords: [
-                [0, 0],
-                [4, 3],
-            ],
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, vectorCongruentRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-
-    it("marks opposite direction as incorrect when match is congruent", () => {
-        // Arrange — reversed direction ⟨-3, -4⟩
-        const guess: PerseusGraphType = {
-            type: "vector",
-            coords: [
-                [3, 4],
-                [0, 0],
-            ],
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, vectorCongruentRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-
-    it("marks wrong magnitude as incorrect when match is congruent", () => {
-        // Arrange — same direction but different length ⟨6, 8⟩
-        const guess: PerseusGraphType = {
-            type: "vector",
-            coords: [
-                [0, 0],
-                [6, 8],
-            ],
-        };
-
-        // Act
-        const result = scoreInteractiveGraph(guess, vectorCongruentRubric);
-
-        // Assert
-        expect(result).toHaveBeenAnsweredIncorrectly();
-    });
-});
